requirements.txt
# Core API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Configuration Management
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LLM Integration
openai>=1.3.0

# SSH Remote Execution
paramiko>=3.3.0

# Async HTTP Client
httpx>=0.25.0

# Logging and Monitoring
structlog>=23.2.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0